app/functions/alert/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-keyvault
azure-identity
urllib3
azure-storage-blob